requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
gunicorn==20.1.0